requirements.txt
customtkinter>=5.2.0
yt-dlp>=2024.01.01
Pillow>=10.0.0
requests>=2.31.0
tkcalendar>=1.6.1
pytest>=7.0.0
